Attempt to fix decade bug 1063: sort image in table header rendered inconsistently.
This code renders the image into a separate TD, and renders all data cells in cells with
colspan=2 for columns that are sortable. In addition the table styles have changed so that
the table itself has cellspacing + cellpadding 0; the columns are rendered with borders 
and the header is more explicitly formatted.